+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- 2019-10-30: ModelFactory::replaceMusclesWithPathActuators() now adds the
+              PathActuators to the Model's ForceSet, and the connectee names
+              for PathPoints are now valid.
+
 - 2019-10-30: Solvers print the date and time before and after solving a 
               problem.